gen/requirements.txt
torch
torchvision
torchaudio
six
espnet
pyopenjtalk==0.1.6
numpy==1.23.5